support_scripts/pyproject.toml
[project]
name = "support-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "polars>=1.35.2",
    "pymediainfo>=7.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
